requirements.txt
absl-py==2.1.0
alembic==1.13.1
annotated-types==0.6.0
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-generator==1.10
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
backcall==0.2.0
batchspawner==1.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
bqplot==0.12.43
cachetools==4.2.4
certifi==2024.2.2
certipy==0.1.3
cffi==1.16.0
charset-normalizer==3.3.2
comm==0.2.1
contourpy==1.1.1
cryptography==42.0.2
cycler==0.12.1
Cython==0.29.14
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.13.1
flatbuffers==2.0.7
fonttools==4.48.1
fqdn==1.5.1
fsspec==2024.2.0
funkify==0.4.5
gast==0.3.3
google-auth==2.15.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.60.1
h11==0.14.0
h5py==2.10.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
imageio==2.34.0
importlib-metadata==7.0.1
importlib-resources==6.1.1
install==1.3.5
ipydatawidgets==4.3.5
ipykernel==6.29.2
ipympl==0.9.3
ipython==8.12.3
ipython-genutils==0.2.0
ipyvolume==0.6.3
ipyvue==1.10.1
ipyvuetify==1.8.10
ipywebrtc==0.6.0
ipywidgets==8.1.2
isoduration==20.11.0
jax==0.4.13
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-client==7.4.1
jupyter-console==6.6.3
jupyter-core==5.7.1
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter-server==1.24.0
jupyter-server-terminals==0.5.2
jupyter-telemetry==0.1.0
jupyterhub==2.3.1
jupyterlab==3.2.0
jupyterlab-pygments==0.3.0
jupyterlab-server==2.25.2
jupyterlab-widgets==3.0.10
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
lazy-loader==0.3
libclang==16.0.6
lxml==5.1.0
Mako==1.3.2
Markdown==3.5.2
MarkupSafe==2.1.5
matplotlib==3.7.4
matplotlib-inline==0.1.6
mistune==3.0.2
ml-dtypes==0.2.0
mpi4py==3.1.3
mpmath==1.3.0
nbclassic==0.3.7
nbclient==0.7.4
nbconvert==7.16.0
nbformat==5.9.2
nest-asyncio==1.6.0
netaddr==0.7.19
netifaces==0.10.9
networkx==3.1
notebook==6.4.13
notebook-shim==0.2.3
numpy==1.21.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oath==1.4.3
oauthlib==3.2.2
opencv-python==4.9.0.80
opt-einsum==3.3.0
overrides==7.7.0
packaging==23.2
pamela==1.1.0
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
pipdeptree==2.15.1
pkgutil-resolve-name==1.3.10
platformdirs==4.2.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
protobuf==3.19.6
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodomex==3.6.1
pygments==2.17.2
pyOpenSSL==24.0.0
pyparsing==3.1.1
pyrsistent==0.20.0
python-dateutil==2.8.2
python-hostlist==1.21
python-json-logger==2.0.7
python-magic==0.4.16
python3-hwloc==2.3.13.0.1.11.9
python3-libnuma==2.3.7
pythreejs==2.4.2
pytz==2024.1
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==25.1.2
qluman==13.4.7
qlunet==13.0.10
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.33.0
requests==2.31.0
requests-oauthlib==1.3.1
requires==0.10.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.17.1
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
scikit-image==0.20.0
scikit-learn==1.3.2
scipy==1.9.1
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.0
stack-data==0.6.3
sympy==1.12
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.7.1
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.34.0
tensorrt==8.6.1.post1
tensorrt-bindings==8.6.1
tensorrt-libs==8.6.1
termcolor==2.4.0
terminado==0.18.0
threadpoolctl==3.2.0
tifffile==2023.7.10
tinycss2==1.2.1
tomli==2.0.1
torch==2.2.0
tornado==6.4
traitlets==5.14.1
traittypes==0.2.1
triton==2.2.0
types-python-dateutil==2.8.19.20240106
typing-extensions==4.5.0
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.0
voila==0.4.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==12.0
werkzeug==3.0.1
widgetsnbextension==4.0.10
wrapt==1.16.0
xtyping==0.8.2
zipp==3.17.0
